use lazy.nvim for plugin manager
