+++ .gitignore
@@ -1,5 +1,4 @@
-vimrcs/my_configs.vim
-luas/my_configs.lua
+vimrcs/my_configs.lua
 
 __pycache__/
 *.py[cod]
