--- luas/plugins.lua
@@ -1,473 +1,0 @@
-------------------------------------------------------------------------------
--- Plugins 
-
-------------------------------------------------------------------------------
--- {{{ nvim-autopairs 
-require("nvim-autopairs").setup {}
-
--- }}}
-
-------------------------------------------------------------------------------
--- {{{ nvim-tree.lua 
-require("nvim-tree").setup({
-    sort_by = "case_sensitive",
-    view = {
-        width = 35,
-    },
-    renderer = {
-        group_empty = true,
-        icons = {
-            show = {
-                file = true,
-                folder = true,
-                folder_arrow = true,
-                git = true,
-                modified = true
-            },
-        },
-    },
-    git = {
-        ignore = false
-    },
-    filters = {
-        dotfiles = false,
-    },
-})
-
-local function open_nvim_tree(data)
-    -- if there're no args, set to current directory
-    if data.file == '' or data.file == nil then
-        data.file = '.'
-    end
-
-    -- buffer is a directory
-    local directory = vim.fn.isdirectory(data.file) == 1
-
-    if not directory then
-        return
-    end
-
-    -- create a new, empty buffer
-    vim.cmd.enew()
-
-    -- wipe the directory buffer
-    vim.cmd.bw(data.buf)
-
-    -- change to the directory
-    vim.cmd.cd(data.file)
-
-    -- open the tree
-    require("nvim-tree.api").tree.open()
-end
-
-vim.api.nvim_create_autocmd({ "VimEnter" }, {
-    callback = open_nvim_tree
-})
-
--- Reload current tree root (navigated by :cd <path>)
-function ReloadNvimTreeRoot()
-    require("nvim-tree.api").tree.change_root(vim.fn.getcwd())
-end
-
-vim.api.nvim_set_keymap("", "<leader>r", ":lua ReloadNvimTreeRoot()<CR>",
-    {noremap = true, silent = true}
-)
-
--- Quickly toggle NvimTree
-vim.api.nvim_set_keymap("", "<leader>t", ":NvimTreeToggle<cr>",
-    {noremap = true, silent = true}
-)
-
-
-------------------------------------------------------------------------------
--- nvim-web-devicons 
-require'nvim-web-devicons'.setup {
-    -- your personnal icons can go here (to override)
-    -- you can specify color or cterm_color instead of specifying both of them
-    -- DevIcon will be appended to `name`
-    override = {
-        zsh = {
-            icon = "",
-            color = "#428850",
-            cterm_color = "65",
-            name = "Zsh"
-        }
-    };
-    -- globally enable different highlight colors per icon (default to true)
-    -- if set to false all icons will have the default icon's color
-    color_icons = true;
-    -- globally enable default icons (default to false)
-    -- will get overriden by `get_icons` option
-    default = true;
-    -- globally enable "strict" selection of icons - icon will be looked up in
-    -- different tables, first by filename, and if not found by extension; this
-    -- prevents cases when file doesn't have any extension but still gets some icon
-    -- because its name happened to match some extension (default to false)
-    strict = true;
-    -- same as `override` but specifically for overrides by filename
-    -- takes effect when `strict` is true
-    override_by_filename = {
-        [".gitignore"] = {
-            icon = "",
-            color = "#f1502f",
-            name = "Gitignore"
-        }
-    };
-    -- same as `override` but specifically for overrides by extension
-    -- takes effect when `strict` is true
-    override_by_extension = {
-        ["log"] = {
-            icon = "",
-            color = "#81e043",
-            name = "Log"
-        }
-    };
-}
-
--- }}}
-
-------------------------------------------------------------------------------
--- {{{ telescope.nvim
-local builtin = require('telescope.builtin')
-vim.keymap.set('n', '<C-p>a', builtin.find_files, {})
-vim.keymap.set('n', '<C-p>', builtin.git_files, {})
-vim.keymap.set('n', '<leader>g', builtin.live_grep, {})
-vim.keymap.set('n', '<leader>b', builtin.buffers, {})
--- }}}
-
-------------------------------------------------------------------------------
--- {{{ nvim-treesitter
-require'nvim-treesitter.configs'.setup {
-    -- A list of parser names, or "all" (the listed parsers should always be installed)
-    ensure_installed = {"c", "lua", "vim", "vimdoc", "python", "cmake", "markdown"},
-
-    -- Install parsers synchronously (only applied to `ensure_installed`)
-    sync_install = false,
-
-    -- Automatically install missing parsers when entering buffer
-    -- Recommendation: set to false if you don't have `tree-sitter` CLI installed locally
-    auto_install = true,
-
-    -- List of parsers to ignore installing (for "all")
-    -- ignore_install = { "javascript" },
-
-    ---- If you need to change the installation directory of the parsers (see -> Advanced Setup)
-    -- parser_install_dir = "/some/path/to/store/parsers", -- Remember to run vim.opt.runtimepath:append("/some/path/to/store/parsers")!
-
-    indent = {
-        enable = true,
-    },
-
-    highlight = {
-        enable = true,
-
-        -- NOTE: these are the names of the parsers and not the filetype. (for example if you want to
-        -- disable highlighting for the `tex` filetype, you need to include `latex` in this list as this is
-        -- the name of the parser)
-        -- list of language that will be disabled
-        -- disable = { "c", "rust" },
-        -- Or use a function for more flexibility, e.g. to disable slow treesitter highlight for large files
-        disable = function(lang, buf)
-            local max_filesize = 100 * 1024 -- 100 KB
-            local ok, stats = pcall(vim.loop.fs_stat, vim.api.nvim_buf_get_name(buf))
-            if ok and stats and stats.size > max_filesize then
-                return true
-            end
-        end,
-
-        -- Setting this to true will run `:h syntax` and tree-sitter at the same time.
-        -- Set this to `true` if you depend on 'syntax' being enabled (like for indentation).
-        -- Using this option may slow down your editor, and you may see some duplicate highlights.
-        -- Instead of true it can also be a list of languages
-        additional_vim_regex_highlighting = false,
-    },
-}
-
--- }}}
-
-------------------------------------------------------------------------------
--- {{{ mason.nvim
-require("mason").setup({
-    ui = {
-        icons = {
-            package_installed = "✓",
-            package_pending = "➜",
-            package_uninstalled = "✗"
-        }
-    }
-})
--- }}}
-
---  {{{ mason-lspconfig
-require("mason-lspconfig").setup({
-    ensure_installed = {"lua_ls", "pyright", "marksman"}
-})
---  }}}
-
-------------------------------------------------------------------------------
--- {{{ nvim-lspconfig 
-local lspconfig = require('lspconfig')
-local capabilities = require('cmp_nvim_lsp').default_capabilities()
-local lsp_defaults = lspconfig.util.default_config
-
-lsp_defaults.capabilities = vim.tbl_deep_extend(
-    'force',
-    lsp_defaults.capabilities,
-    capabilities
-)
-
--- pyright
-local python_path = vim.fn.trim(vim.fn.system("which python"))
-lspconfig.pyright.setup {
-    settings = {
-        python = {
-            pythonPath = python_path,
-        }
-    },
-    capabilities = capabilities
-}
-
--- tsserver
-lspconfig.tsserver.setup {
-    capabilities = capabilities
-}
-
--- rust_analyzer
-lspconfig.rust_analyzer.setup {
-    -- Server-specific settings. See `:help lspconfig-setup`
-    settings = {
-        ['rust-analyzer'] = {
-            diagnostic = {
-                enable = false;
-            }
-        },
-    },
-    capabilities = capabilities
-}
-
--- ccls
--- lspconfig.ccls.setup {  
---     capabilities = capabilities
--- }
-
--- clangd.
-lspconfig.clangd.setup {
-    capabilities = capabilities;
-    support_single_file = true;
-}
-
---cmake
-lspconfig.cmake.setup{}
-
---lua for nvim
-lspconfig.lua_ls.setup {
-  on_init = function(client)
-    local path = client.workspace_folders[1].name
-    if not vim.loop.fs_stat(path..'/.luarc.json') and not vim.loop.fs_stat(path..'/.luarc.jsonc') then
-      client.config.settings = vim.tbl_deep_extend('force', client.config.settings, {
-        Lua = {
-          runtime = {
-            -- Tell the language server which version of Lua you're using
-            -- (most likely LuaJIT in the case of Neovim)
-            version = 'LuaJIT'
-          },
-          -- Make the server aware of Neovim runtime files
-          workspace = {
-            checkThirdParty = false,
-            library = {
-              vim.env.VIMRUNTIME
-              -- "${3rd}/luv/library"
-              -- "${3rd}/busted/library",
-            }
-            -- or pull in all of 'runtimepath'. NOTE: this is a lot slower
-            -- library = vim.api.nvim_get_runtime_file("", true)
-          }
-        }
-      })
-
-      client.notify("workspace/didChangeConfiguration", { settings = client.config.settings })
-    end
-    return true
-  end
-}
-
--- marksman
-lspconfig.marksman.setup{}
-
--- Global mappings.
--- See `:help vim.diagnostic.*` for documentation on any of the below functions
-vim.keymap.set('n', '<space>e', vim.diagnostic.open_float)
-vim.keymap.set('n', '[d', vim.diagnostic.goto_prev)
-vim.keymap.set('n', ']d', vim.diagnostic.goto_next)
-vim.keymap.set('n', '<space>q', vim.diagnostic.setloclist)
-
--- Use LspAttach autocommand to only map the following keys
--- after the language server attaches to the current buffer
-vim.api.nvim_create_autocmd('LspAttach', {
-    group = vim.api.nvim_create_augroup('UserLspConfig', {}),
-    callback = function(ev)
-        -- Enable completion triggered by <c-x><c-o>
-        vim.bo[ev.buf].omnifunc = 'v:lua.vim.lsp.omnifunc'
-
-        -- Buffer local mappings.
-        -- See `:help vim.lsp.*` for documentation on any of the below functions
-        local opts = { buffer = ev.buf }
-        vim.keymap.set('n', 'gD', vim.lsp.buf.declaration, opts)
-        vim.keymap.set('n', 'gd', vim.lsp.buf.definition, opts)
-        vim.keymap.set('n', 'K', vim.lsp.buf.hover, opts)
-        vim.keymap.set('n', 'gi', vim.lsp.buf.implementation, opts)
-        vim.keymap.set('n', '<C-k>', vim.lsp.buf.signature_help, opts)
-        vim.keymap.set('n', '<space>wa', vim.lsp.buf.add_workspace_folder, opts)
-        vim.keymap.set('n', '<space>wr', vim.lsp.buf.remove_workspace_folder, opts)
-        vim.keymap.set('n', '<space>wl', function()
-            print(vim.inspect(vim.lsp.buf.list_workspace_folders()))
-        end, opts)
-        vim.keymap.set('n', '<space>D', vim.lsp.buf.type_definition, opts)
-        vim.keymap.set('n', '<space>rn', vim.lsp.buf.rename, opts)
-        vim.keymap.set({ 'n', 'v' }, '<space>ca', vim.lsp.buf.code_action, opts)
-        vim.keymap.set('n', 'gr', vim.lsp.buf.references, opts)
-        vim.keymap.set('n', '<space>f', function()
-            vim.lsp.buf.format { async = true }
-        end, opts)
-    end,
-})
-
--- }}}
-
-------------------------------------------------------------------------------
--- {{{ nvim-cmp
--- Set up nvim-cmp.
-local cmp = require'cmp'
-
-cmp.setup({
-    -- to remove LSP priority to select completion
-    preselect = cmp.PreselectMode.None,
-    snippet = {
-        -- REQUIRED - you must specify a snippet engine
-        expand = function(args)
-            vim.fn["vsnip#anonymous"](args.body) -- For `vsnip` users.
-            -- require('luasnip').lsp_expand(args.body) -- For `luasnip` users.
-            -- require('snippy').expand_snippet(args.body) -- For `snippy` users.
-            -- vim.fn["UltiSnips#Anon"](args.body) -- For `ultisnips` users.
-        end,
-    },
-    window = {
-        -- completion = cmp.config.window.bordered(),
-        -- documentation = cmp.config.window.bordered(),
-    },
-    mapping = cmp.mapping.preset.insert({
-        ['<C-b>'] = cmp.mapping.scroll_docs(-4),
-        ['<C-f>'] = cmp.mapping.scroll_docs(4),
-        ['<C-Space>'] = cmp.mapping.complete(),
-        ['<C-e>'] = cmp.mapping.abort(),
-        ['<CR>'] = cmp.mapping.confirm({ select = true }), -- Accept currently selected item. Set `select` to `false` to only confirm explicitly selected items.
-    }),
-    sources = cmp.config.sources({
-        { name = 'nvim_lsp' },
-        { name = 'vsnip' }, -- For vsnip users.
-        -- { name = 'luasnip' }, -- For luasnip users.
-        -- { name = 'ultisnips' }, -- For ultisnips users.
-        -- { name = 'snippy' }, -- For snippy users.
-    }, {
-            { name = 'buffer' },
-        })
-})
-
--- Set configuration for specific filetype.
-cmp.setup.filetype('gitcommit', {
-    sources = cmp.config.sources({
-        { name = 'git' }, -- You can specify the `git` source if [you were installed it](https://github.com/petertriho/cmp-git).
-    }, {
-            { name = 'buffer' },
-        })
-})
-
--- Use buffer source for `/` and `?` (if you enabled `native_menu`, this won't work anymore).
-cmp.setup.cmdline({ '/', '?' }, {
-    mapping = cmp.mapping.preset.cmdline(),
-    sources = {
-        { name = 'buffer' }
-    }
-})
-
--- Use cmdline & path source for ':' (if you enabled `native_menu`, this won't work anymore).
-cmp.setup.cmdline(':', {
-    mapping = cmp.mapping.preset.cmdline(),
-    sources = cmp.config.sources({
-        { name = 'path' }
-    }, {
-            { name = 'cmdline' }
-        })
-})
-
-cmp.setup({
-    mapping = {
-        -- If nothing is selected, <CR> will make a new line
-        ["<CR>"] = cmp.mapping({
-            i = function(fallback)
-                if cmp.visible() and cmp.get_active_entry() then
-                    cmp.confirm({ behavior = cmp.ConfirmBehavior.Replace, select = false })
-                else
-                    fallback()
-                end
-            end,
-        }),
-        -- Tab to go to next suggestion
-        ["<Tab>"] = cmp.mapping({
-            c = function(fallback)
-                if cmp.visible() then
-                    cmp.select_next_item({ behavior = cmp.SelectBehavior.Insert })
-                else
-                    fallback()
-                end
-            end,
-            i = function(fallback)
-                if cmp.visible() then
-                    cmp.select_next_item({ behavior = cmp.SelectBehavior.Insert })
-                else
-                    fallback()
-                end
-            end,
-            s = function(fallback)
-                if cmp.visible() then
-                    cmp.select_next_item({ behavior = cmp.SelectBehavior.Insert })
-                else
-                    fallback()
-                end
-            end,
-        }),
-        -- S-Tab to go to last suggestion
-        ["<S-Tab>"] = cmp.mapping({
-            c = function(fallback)
-                if cmp.visible() then
-                    cmp.select_prev_item({ behavior = cmp.SelectBehavior.Insert })
-                else
-                    fallback()
-                end
-            end,
-            i = function(fallback)
-                if cmp.visible() then
-                    cmp.select_prev_item({ behavior = cmp.SelectBehavior.Insert })
-                else
-                    fallback()
-                end
-            end,
-            s = function(fallback)
-                if cmp.visible() then
-                    cmp.select_prev_item({ behavior = cmp.SelectBehavior.Insert })
-                else
-                    fallback()
-                end
-            end,
-        }),
-    }
-})
-
--- If you want insert `(` after select function or method item
-local cmp_autopairs = require('nvim-autopairs.completion.cmp')
-cmp.event:on(
-    'confirm_done',
-    cmp_autopairs.on_confirm_done()
-)
-
--- }}}
