+++ my_plugins/.gitignore
@@ -1,1 +1,1 @@
-**/
+*
