+++ plugins/git.lua
@@ -1,0 +1,10 @@
+return {
+    --------------------------------------------------------------------------
+    -- {{{ vim-fugitive
+    {
+        "tpope/vim-fugitive",
+        lazy = true,
+        cmd = "Git"
+    },
+    -- }}}
+}
