+++ vimrcs/init.lua
@@ -1,12 +1,38 @@
 -- INDEX
+-- => Lazy.nvim for plugin manager
 -- => General
 -- => UI
 -- => Apperance
 -- => Mapping
 -- => Autocmd
+-- => Misc
 
 ------------------------------------------------------------------------------
--- {{{ => General 
+-- {{{ => Lazy.nvim
+--
+-- Set Leader key to Space (must be put before Lazy.nvim)
+vim.g.mapleader = " "
+
+-- Install Lazy.nvim
+local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
+if not vim.loop.fs_stat(lazypath) then
+    vim.fn.system({
+        "git",
+        "clone",
+        "--filter=blob:none",
+        "https://github.com/folke/lazy.nvim.git",
+        "--branch=stable", -- latest stable release
+        lazypath,
+    })
+end
+vim.opt.rtp:prepend(lazypath)
+
+-- import all plugins from "plugins/" directory
+require("lazy").setup("plugins")
+-- }}}
+
+------------------------------------------------------------------------------
+-- {{{ => General
 -- Number of lines in history
 vim.opt.history = 500
 
@@ -41,8 +67,7 @@ vim.opt.wrap = true  -- Wrap lines
 vim.opt.relativenumber = true
 
 -- Enable filetype plugins
-vim.api.nvim_command("filetype plugin on")
-vim.api.nvim_command("filetype indent on")
+vim.api.nvim_command("filetype plugin indent on")
 
 -- Use magic for regex
 vim.opt.magic = true
@@ -64,7 +89,6 @@ vim.g.os = vim.loop.os_uname().sysname
 
 -- Get ~ directory path
 HOME = os.getenv("HOME")
-
 --}}}
 
 ------------------------------------------------------------------------------
@@ -126,97 +150,6 @@ vim.cmd("syntax enable")
 -- Background
 vim.opt.background = "dark"
 
--- Colorscheme
----- catppuccin
-require("catppuccin").setup({
-    flavour = "mocha", -- latte, frappe, macchiato, mocha
-    background = { -- :h background
-        light = "latte",
-        dark = "mocha",
-    },
-    transparent_background = false, -- disables setting the background color.
-    show_end_of_buffer = false, -- shows the '~' characters after the end of buffers
-    term_colors = false, -- sets terminal colors (e.g. `g:terminal_color_0`)
-    dim_inactive = {
-        enabled = false, -- dims the background color of inactive window
-        shade = "dark",
-        percentage = 0.15, -- percentage of the shade to apply to the inactive window
-    },
-    no_italic = false, -- Force no italic
-    no_bold = false, -- Force no bold
-    no_underline = false, -- Force no underline
-    styles = { -- Handles the styles of general hi groups (see `:h highlight-args`):
-        comments = { "italic" }, -- Change the style of comments
-        conditionals = { "italic" },
-        loops = {},
-        functions = {},
-        keywords = {},
-        strings = {},
-        variables = {},
-        numbers = {},
-        booleans = {},
-        properties = {},
-        types = {},
-        operators = {},
-    },
-    color_overrides = {},
-    custom_highlights = {},
-    integrations = {
-        cmp = true,
-        nvimtree = true,
-        treesitter = true,
-        mason = true,
-        native_lsp = {
-            enabled = true,
-            virtual_text = {
-                errors = { "italic" },
-                hints = { "italic" },
-                warnings = { "italic" },
-                information = { "italic" },
-            },
-            underlines = {
-                errors = { "underline" },
-                hints = { "underline" },
-                warnings = { "underline" },
-                information = { "underline" },
-            },
-            inlay_hints = {
-                background = true,
-            },
-        },
-    }
-})
-
----- rose-pine
-require('rose-pine').setup({
-	--- @usage 'auto'|'main'|'moon'|'dawn'
-	variant = 'auto',
-	--- @usage 'main'|'moon'|'dawn'
-	dark_variant = 'main',
-	bold_vert_split = false,
-	dim_nc_background = false,
-	disable_background = false,
-	disable_float_background = false,
-	disable_italics = true,
-
-	-- Change specific vim highlight groups
-	-- https://github.com/rose-pine/neovim/wiki/Recipes
-	highlight_groups = {
-		ColorColumn = { bg = 'rose' },
-
-		-- Blend colours against the "base" background
-		CursorLine = { bg = 'foam', blend = 10 },
-		StatusLine = { fg = 'love', bg = 'love', blend = 10 },
-
-		-- By default each group adds to the existing config.
-		-- If you only want to set what is written in this config exactly,
-		-- you can set the inherit option:
-		Search = { bg = 'gold', inherit = false },
-	}
-})
-
----- set colorscheme
-vim.cmd.colorscheme("catppuccin-macchiato")
 
 -- Enable 256 colors palette in Gnome Terminal
 if os.getenv("COLORTERM") == "gnome-terminal" then
@@ -247,12 +180,13 @@ function SeeWaifu(color)
     vim.api.nvim_set_hl(0, "TelescopeNormal", {bg = "none"})
 end
 
+-- Set colorscheme
+vim.cmd.colorscheme("catppuccin-macchiato")
+
 -- }}}
 
 ------------------------------------------------------------------------------
 -- {{{ => Mapping
--- Set Leader key to Space
-vim.g.mapleader = " "
 
 -- Quickly edit vimrc file in the left window pane
 vim.api.nvim_set_keymap("n", "<leader>ev", "<c-w>l:e $MYVIMRC<cr>",
@@ -415,4 +349,48 @@ vim.api.nvim_create_autocmd("FileType", {
     command = [[setlocal foldmethod=marker]]
 })
 
+-- }}}
+
+------------------------------------------------------------------------------
+-- {{{ => Misc
+-- netrw
+------------------------------------------------------------------------------
+vim.cmd([[
+" Disables netrw by default
+let g:loaded_netrw=1
+let g:loaded_netrwPlugin=1
+
+" Hit enter in the file browser to open the selected
+" file with :vsplit to the right of the browser.
+let g:netrw_browse_split=4
+let g:netrw_altv=1
+
+" Default to tree mode
+let g:netrw_liststyle=3
+
+" Keep the current directory and the browsing directory synced
+let g:netrw_keepdir=0
+
+" Window size
+let g:netrw_winsize=20
+
+" Hide banner
+let g:netrw_banner=0
+
+" Enable recursive copy
+let g:netrw_localcopydircmd='cp -r'
+
+" Sort is affecting only: directories on the top, files below
+let g:netrw_sort_sequence = '[\/]$,*'
+
+" Highlight marked files
+hi! link netrwMarkFile Search
+]])
+
+-- the silver searcher
+vim.cmd([[
+if executable('ag')
+    let g:ackprg = 'ag --vimgrep'
+endif
+]])
 -- }}}
